chore: test the mocks policy
